fix: Rename gate_model_simulator.py to simulator.py
